Organização do codigo e toggle do SSAO
organização entre muitas aspas
