+++ sourcecode/02_hello_triangle/02_hellotriangle.vcxproj.filters
@@ -25,6 +25,7 @@
     <ClCompile Include="Mesh.cpp" />
     <ClCompile Include="Shader.cpp" />
     <ClCompile Include="Model.cpp" />
+    <ClCompile Include="OpenGLApp.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="gl_utils.h">
@@ -37,12 +38,28 @@
     <ClInclude Include="Mesh.h" />
     <ClInclude Include="Shader.h" />
     <ClInclude Include="Model.h" />
+    <ClInclude Include="OpenGLApp.h" />
   </ItemGroup>
   <ItemGroup>
-    <None Include="test_fs.glsl">
+    <None Include="ssao_blur_FS.glsl">
       <Filter>Shaders</Filter>
     </None>
-    <None Include="test_vs.glsl">
+    <None Include="ssao_VS.glsl">
+      <Filter>Shaders</Filter>
+    </None>
+    <None Include="ssao_lighting_FS.glsl">
+      <Filter>Shaders</Filter>
+    </None>
+    <None Include="ssao_geometryCube_FS.glsl">
+      <Filter>Shaders</Filter>
+    </None>
+    <None Include="ssao_geometry_VS.glsl">
+      <Filter>Shaders</Filter>
+    </None>
+    <None Include="ssao_geometry_FS.glsl">
+      <Filter>Shaders</Filter>
+    </None>
+    <None Include="ssao_FS.glsl">
       <Filter>Shaders</Filter>
     </None>
   </ItemGroup>
