+++ sourcecode/02_hello_triangle/Camera.cpp
@@ -1,12 +1,6 @@
 #include "Camera.h"
 
 
-
-Camera::Camera()
-{
-}
-
-
 Camera::~Camera()
 {
 }
