+++ sourcecode/02_hello_triangle/Camera.h
@@ -43,7 +43,6 @@ public:
 	float MouseSensitivity;
 	float Zoom;
 
-	Camera();
 	~Camera();
 	// Constructor with vectors
 	Camera(glm::vec3 position = glm::vec3(0.0f, 0.0f, 0.0f), glm::vec3 up = glm::vec3(0.0f, 1.0f, 0.0f), float yaw = YAW, float pitch = PITCH) : Front(glm::vec3(0.0f, 0.0f, -1.0f)), MovementSpeed(SPEED), MouseSensitivity(SENSITIVITY), Zoom(ZOOM)
