+++ sourcecode/02_hello_triangle/nanosuit/LICENSE.txt
@@ -1,0 +1,6 @@
+Original Nanosuit model by ForrestPL:
+http://tf3dm.com/3d-model/crysis-2-nanosuit-2-97837.html
+
+Slightly modified for use in the LearnOpenGL.com tutorials (by Joey de Vries)
+
+For personal use only
