--- sourcecode/02_hello_triangle/test_fs.glsl
@@ -1,12 +1,0 @@
-#version 410
-
-out vec4 FragColor;
-
-in vec2 TexCoords;
-
-uniform sampler2D texture_diffuse1;
-
-void main()
-{    
-    FragColor = texture(texture_diffuse1, TexCoords);
-}
